Extract image pipeline to integrations/image_pipeline.py; fix DST and vision model
- Move _vision_call, _triage_image, _get_meal_label, all per-type handlers,
  _try_create_calendar_event, and last_event_cache out of bot.py into
  integrations/image_pipeline.py so they are independently testable
- Add public route_image(image_b64, caption, user_id) -> (str, str) entry point
- Fix DST bug in _get_meal_label: replace hardcoded UTC-5 with ZoneInfo("America/New_York")
- Fix vision model name: llama-3.2-11b-vision-preview -> llama-3.2-11b-vision-instruct
- Update bot.py to import route_image, _try_create_calendar_event, last_event_cache
  from image_pipeline; handle_photo now calls route_image in one line
